requirements.txt
nautilus_trader==1.202.0
qoery>=0.6.0
python-dotenv>=1.0.0
pandas>=2.0.0